Started working on a proper GC, clean of the code so the alignment will be passed around properly

+++ src/util/stb_ds.h
@@ -1071,7 +1071,7 @@ static size_t stbds_siphash_bytes(void *p, size_t len, size_t seed)
     } while (0)
 
   for (i=0; i+sizeof(size_t) <= len; i += sizeof(size_t), d += sizeof(size_t)) {
-    data = d[0] | (d[1] << 8) | (d[2] << 16) | (d[3] << 24);
+    data = d[0] | (d[1] << 8) | (d[2] << 16) | ((uint32_t)d[3] << 24);
     data |= (size_t) (d[4] | (d[5] << 8) | (d[6] << 16) | ((uint32_t)d[7] << 24)) << 16 << 16; // discarded if size_t == 4
 
     v3 ^= data;
